deep freeze data in test mode
Therefore, an error will be raised any time the app tries to mutate data, giving us a
warning and hence protection against mutative behavior.  This also frees use from using
ImmutableJS for solely the purpose of protecting against mutation; now we are free to
just choose ImmutableJS structures when there is a functional or performance benefit.